Cryptograph with vector initialization
